+++ src/main/java/com/team/demo/DemoApplication.java
@@ -3,10 +3,11 @@ package com.team.demo;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.ServletComponentScan;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-/*@ComponentScan(basePackages = {"com.team.demo.config"})*/
+@ServletComponentScan(basePackages = {"com.team.demo.generator.controller.jwt"})
 public class DemoApplication {
 
 	public static void main(String[] args) {
